Move page components
