+++ coinpurse.client/src/containers/App.tsx
@@ -1,6 +1,6 @@
 import { Routes, Route } from "react-router";
-import DashboardPage from "./DashboardPage";
-import SettingsPage from "./SettingsPage";
+import DashboardPage from "./Pages/DashboardPage";
+import SettingsPage from "./Pages/SettingsPage";
 import AppLayout from "./AppLayout";
 import { Box } from "@mui/material";
