+++ coinpurse.client/src/containers/Pages/DashboardPage.tsx
