+++ coinpurse.client/src/containers/Pages/SettingsPage.tsx
